requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
langserve
python-dotenv
openai
requests
beautifulsoup4
pypdf
chromadb
faiss-cpu
sentence-transformers
sse-starlette
streamlit
bs4
pypdf
chromadb
faiss-cpu
wikipedia
arxiv
langchain-groq